+++ source/input/actions.hpp
@@ -8,14 +8,14 @@ namespace eru
    struct ValueAction final
    {
       std::unordered_set<Input> inputs{};
-      double deadzone{ 0.25 };
+      float deadzone{ 0.25 };
    };
 
    struct AxisAction final
    {
       std::unordered_set<Input> positive_inputs{};
       std::unordered_set<Input> negative_inputs{};
-      double deadzone{ 0.25 };
+      float deadzone{ 0.25 };
    };
 
    struct VectorAction final
@@ -24,7 +24,7 @@ namespace eru
       std::unordered_set<Input> negative_x_inputs{};
       std::unordered_set<Input> positive_y_inputs{};
       std::unordered_set<Input> negative_y_inputs{};
-      double deadzone{ 0.25 };
+      float deadzone{ 0.25 };
    };
 }
